requirements.txt
# chatlm runtime requirements.
# Pinned to actually-tested versions so a fresh install is reproducible.
# See docs in this file for the rationale behind each non-trivial pin.

# --- web stack ----------------------------------------------------------
fastapi==0.136.1
# uvicorn 0.32 is older than upstream (0.46) but works fine here; we hold
# back to avoid breaking-API churn until there's a concrete reason to bump.
uvicorn[standard]==0.32.0
starlette==1.0.0
httpx==0.28.1
pydantic==2.13.3
pydantic-settings==2.6.0

# --- system / utility ---------------------------------------------------
psutil==6.1.0

# --- vision pipelines ---------------------------------------------------
ultralytics==8.3.40
opencv-python-headless==4.10.0.84
pillow==11.0.0

# --- HF ecosystem -------------------------------------------------------
# transformers 5.x is the current major; we run on 5.6.x for the latest
# Qwen3.5/FLUX.2/SD3.5-Large support paths.
transformers==5.6.2
diffusers==0.37.1
huggingface-hub==1.12.0
accelerate==1.13.0
peft==0.19.1
safetensors==0.7.0
sentencepiece==0.2.1

# --- Apple Silicon native (mlx) ----------------------------------------
# mlx-vlm pulls mlx + mlx-lm transitively; we pin them explicitly so a
# fresh venv on a new mac doesn't surprise-jump a major.
mlx==0.31.2; sys_platform == "darwin" and platform_machine == "arm64"
mlx-lm==0.31.3; sys_platform == "darwin" and platform_machine == "arm64"
mlx-vlm; sys_platform == "darwin" and platform_machine == "arm64"

# NOTE: torch is pulled in transitively by ultralytics. For NVIDIA CUDA, install
# the matching wheel first:  pip install torch --index-url https://download.pytorch.org/whl/cu121
# For macOS (Apple Silicon), the default PyPI wheel already includes MPS support.
